pyproject.toml
[tool.poetry]
name = "discord_bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]


[tool.poetry.dependencies]
python = "^3.10"
loguru = "^0.6.0"
"discord.py" = "^1.7.3"


[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
black = "^22.1.0"
isort = "^5.10.1"
mypy = "^0.942"
pyproject-flake8 = "^0.0.1-alpha.2"
flake8-bugbear = "^22.3.23"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.black]
target-version = ['py310']
line-length = 120


[tool.isort]
line_length = 120
multi_line_output = 3
include_trailing_comma = true
known_local_folder = ['config', ]


[tool.flake8]
max-line-length = 120
max-complexity = 10
ignore = "E203,E266,W503,"


[tool.mypy]
python_version = "3.10"
no-strict_optional = true
ignore_missing_imports = true
check_untyped_defs = true
